refactor: seperate session managing function
